cloud function implemented and post shopping moved to data.js
